segmently_django/requirements.txt
aniso8601==3.0.2
anyjson==0.3.3
autoenv==1.0.0
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
dj-database-url==0.4.1
dj-static==0.0.6
Django==2.2.3
django-extensions==1.7.9
django-filter==2.0.0
django-rest-swagger==2.1.2
django-reversion==3.0.2
django-toolbelt==0.0.1
djangorestframework==3.9
djangorestframework-filters==1.0.0.dev0
djongo==1.2.33
dry-rest-permissions==0.1.10
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==18.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
olefile==0.44
openapi-codec==1.3.2
promise==2.2.1
pymongo==3.8.0
pytz==2017.3
PyYAML==3.13
requests==2.18.4
Rx==1.6.1
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
typing==3.6.6
uritemplate==3.0.0
urllib3==1.22